+++ grails-app/services/io/xh/hoist/monitor/MemoryMonitoringService.groovy
@@ -13,6 +13,8 @@ import io.xh.hoist.util.Timer
 import java.util.concurrent.ConcurrentHashMap
 
 import static io.xh.hoist.util.DateTimeUtils.SECONDS
+import static io.xh.hoist.util.DateTimeUtils.HOURS
+import static io.xh.hoist.util.DateTimeUtils.intervalElapsed
 import static java.lang.Runtime.getRuntime
 
 /**
@@ -23,6 +25,7 @@ class MemoryMonitoringService extends BaseService {
 
     private Map<Long, Map> _snapshots = new ConcurrentHashMap()
     private Timer _snapshotTimer
+    private Date _lastInfoLogged
 
     void init() {
         _snapshotTimer = createTimer(
@@ -44,13 +47,25 @@ class MemoryMonitoringService extends BaseService {
      */
     Map takeSnapshot() {
         def newSnap = getStats()
-        _snapshots.put(System.currentTimeMillis(), newSnap)
+
+        _snapshots[System.currentTimeMillis()] = newSnap
 
         // Don't allow snapshot history to grow endlessly - cap @ 1440 samples, i.e. 24 hours of
         // history if left at default config interval of one snap/minute.
         if (_snapshots.size() > 1440) {
             def oldest = _snapshots.keys().toList().min()
             _snapshots.remove(oldest)
+        }
+
+        def msg = logMessagesForSnapshot(newSnap)
+        if (newSnap.usedPctTotal > 90) {
+            logWarn(msg)
+            logWarn("MEMORY USAGE ABOVE 90%")
+        } else if (intervalElapsed(1 * HOURS, _lastInfoLogged)) {
+            logInfo(msg)
+            _lastInfoLogged = new Date()
+        } else {
+            logDebug(msg)
         }
 
         return newSnap
@@ -93,6 +108,17 @@ class MemoryMonitoringService extends BaseService {
             usedPctTotal: round((used * 100) / total),
             totalPctMax: round((total * 100) / max)
         ]
+    }
+
+    private List logMessagesForSnapshot(Map snap) {
+        def totalHeap = snap.totalHeapMb,
+            maxHeap = snap.maxHeapMb,
+            usedHeap = snap.usedHeapMb,
+            freeHeap = snap.freeHeapMb,
+            usedPctTotal = snap.usedPctTotal
+
+        return ["Total=${totalHeap}MB", "Max=${maxHeap}MB", "Used=${usedHeap}MB", "Free=${freeHeap}MB",
+                "Used Percent of Total=${usedPctTotal}%"]
     }
 
     void clearCaches() {