+++ CHANGELOG.md
@@ -9,6 +9,11 @@
   is required.
 * `DateTimeUtils` app/server timezone conversion utils default to current day/date if called without arguments.
 
+### 🎁 New Features
+
+* New logging in `MemoryMonitoringService`.
+
+
 [Commit Log](https://github.com/xh/hoist-core/compare/v14.0.0..v14.0.1)
 
 ## 14.0.0 - 2022-07-12